requirement.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==5.1.0
Django==3.2.25
django-allauth==0.61.1
django-filter==23.5
django-model-utils==4.3.1
djangorestframework==3.15.1
idna==3.10
importlib-metadata==6.7.0
Markdown==3.4.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg==3.1.20
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2025.2
urllib3==2.0.7
zipp==3.15.0
